Convert project to Maven
It's been a long time coming, but I'm finally
replacing Gradle with Maven. Gradle has too
many shortcomings, mainly breaking backwards
compatibility with existing configurations on
upgrades, that I will choose to use a more tried
and true system. None of the fancy functionalities
offered by Gradle have any application in this
simple project, and with the limited time I have
to devote to this does not justify the effort.